+++ MDL/BindModel.cs
@@ -12,7 +12,6 @@ namespace MDL
     {
         public Article _Article { get; set; }
         public Category _Category { get; set; }
-        public FileModel _File { get; set; }
     }
 
     public class Article
@@ -30,15 +29,5 @@ namespace MDL
         public string CategoryName { get; set; }
     }
 
-    public class FileModel
-    {
-        public int ArticleId { get; set; }
-        public int FileId { get; set; }
-        public string FilePath { get; set; }
-
-        [Required(ErrorMessage = "Selecteer een bestand.")]
-        [Display(Name = "Selecteer een bestand(en)")]
-        public HttpPostedFileBase[] Files { get; set; }
-
-    }
+    
 }
